+++ ringbuffer.h
@@ -15,6 +15,6 @@ public:
 	RingBuffer();
 	~RingBuffer();
 	void add(const ap_uint<CHUNK_SIZE_BITS> *i_fragment);
-	const void get(const uint16_t i_index, ap_uint<CHUNK_SIZE_BITS> *o_fragment);
+	const void get(const uint64_t i_index, ap_uint<CHUNK_SIZE_BITS> *o_fragment);
 	void getCurrentIndex(uint32_t *o_index);
 };
